refactor: middleware
