+++ apps/www/src/api/middlewares/withRateLimiter.ts
@@ -1,72 +1,23 @@
 // TODO: add this to the next package or maybe make it for all fetcher serverless things
 // or just inject the resut and shit
-import type { TsRestRequest } from "@ts-rest/serverless/next";
-
-import { tsr } from "@ts-rest/serverless/next";
-import type { Contract } from "~/api/contract";
-import type { TsrContext } from "~/api/context";
-import type { Keys } from "ts-roids";
-import type { NextRequest } from "next/server";
 import { logger } from "@ashgw/observability";
+import { createMiddleware } from "~/api/middleware";
+import { middlewareFn } from "~/api/middleware";
+import type { ContractRoute } from "~/api/middleware";
 
-type CRoute = Contract[Keys<Contract>];
-
-type MergeTsrContextWith<C> = TsrContext & {
-  ctx: TsrContext["ctx"] & C;
-};
-
-interface MiddlewareRespone {
-  nextRequest: NextRequest;
-}
-
-type MiddlewareRequest<LocalCtx> = TsRestRequest &
-  TsrContext &
-  MergeTsrContextWith<LocalCtx>;
-
-type MiddlewareFn<LocalCtx> = (
-  req: MiddlewareRequest<LocalCtx>,
-  res: MiddlewareRespone,
-) => unknown;
-
-export function middlewareFn<LocalCtx extends object>(
-  fn: MiddlewareFn<LocalCtx>,
-) {
-  return (req: MiddlewareRequest<LocalCtx>) => {
-    fn(req, { nextRequest: req as unknown as NextRequest });
-  };
-}
-
-export function createMiddleware<R extends CRoute, LocalCtx extends object>({
-  route,
-  middlewareFn,
-}: {
-  route: R;
-  middlewareFn: MiddlewareFn<LocalCtx>;
-}) {
-  const build = tsr.routeWithMiddleware(route)<
-    TsrContext,
-    MergeTsrContextWith<LocalCtx>
-  >;
-
-  type BuildOpts = Parameters<typeof build>[0];
-
-  return (handler: BuildOpts["handler"]) => {
-    return build({
-      handler,
-      middleware: [middlewareFn],
-    });
-  };
-}
-
-interface RateLimiter {
+interface RateLimiterCtx {
   put: () => unknown;
   pop: () => unknown;
 }
 
-export function withRateLimiter<R extends CRoute>({ route }: { route: R }) {
-  return createMiddleware<R, RateLimiter>({
+export function withRateLimiter<R extends ContractRoute>({
+  route,
+}: {
+  route: R;
+}) {
+  return createMiddleware<R, RateLimiterCtx>({
     route,
-    middlewareFn: middlewareFn<RateLimiter>((req, res) => {
+    middlewareFn: middlewareFn<RateLimiterCtx>((req, res) => {
       logger.log(req.ctx.pop);
       logger.log(req.ctx.requestedAt);
       logger.log(res.nextRequest);
